+++ source/pza/core/Attribute.cxx
@@ -1,6 +1,8 @@
-#include <core/Attribute.hxx>
+#include <pza/core/Attribute.hxx>
 
 #include <iostream>
+
+using namespace pza;
 
 void Attribute::onMessage(const std::string &topic, const std::string &payload)
 {
@@ -8,7 +10,7 @@ void Attribute::onMessage(const std::string &topic, const std::string &payload)
 
     spdlog::trace("Attribute {:s} received message on topic {:s}, data : {:s}", _name, topic, payload);
 
-    if (Utils::Json::ParseJson(payload, data) == -1)
+    if (utils::json::ParseJson(payload, data) == -1)
         return ;
     onData(data[_name]);
 }