+++ source/pza/interfaces/Psu.cxx
@@ -1,4 +1,6 @@
-#include <interfaces/Psu.hxx>
+#include <pza/interfaces/Psu.hxx>
+
+using namespace pza;
 
 Psu::Psu(Client &client, const std::string &name)
     : Interface(client, name)