+++ CHANGELOG.md
@@ -1,4 +1,14 @@
+# [0.1.0]
+## Added
+- Core, Client and Interface
+- Psu interface
+- Conan support
+- Google test
+- cppcheck
+- Debug and Release build versions
+## Removed
+- cmake fetch_content
 # [0.0.0]
-##Added
+## Added
 - CMake architecture done
 - Dependencies added to the build